modify portfolio based on advice and change layout based on bootstrap
