--- portfolio/src/main/java/com/google/sps/servlets/MyServlet.java
@@ -1,35 +1,0 @@
-package com.google.sps.servlets;
-
-import java.io.IOException;
-import java.util.ArrayList;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import com.google.gson.Gson;
-
-@WebServlet("/information")
-public final class MyServlet extends HttpServlet {
-  @Override
-  public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
-    // Convert the server info to JSON
-    ArrayList<String> info = new ArrayList<>();
-    info.add("I enjoy basketball, badmiton, pingpong, and running");
-    info.add("I will learn tennies and boxing this summer");
-    info.add("I am learning stakeboarding now");
-    String json = convertToJson(info);
-
-    // Send the JSON as the response
-    response.setContentType("application/json;");
-    response.getWriter().println(json);
-  }
-
-  /**
-   * Converts a ServerStats instance into a JSON string using manual String concatentation.
-   */
-  private String convertToJson(ArrayList<String> list) {
-    Gson gson = new Gson();
-    String json = gson.toJson(list);
-    return json;
-  }
-}
